实现基本绘制功能

+++ main.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
+import time
+start_time = time.perf_counter()
 import sys
 import os
-import time
 import re
 import logging
 import json
@@ -10,20 +11,22 @@ from pathlib import Path
 import asyncio
 
 import numpy as np
-from PySide6.QtCore import Qt, QSize,QEvent,QCoreApplication,QPointF
-from PySide6.QtWidgets import QApplication, QMainWindow, QLabel, QVBoxLayout, QWidget,QMessageBox,QWidgetItem,QFormLayout,QToolButton,QSpinBox,QDoubleSpinBox
-from PySide6.QtGui import QIcon,QColor,QSurfaceFormat,QKeyEvent,QMouseEvent
+from PySide6.QtCore import Qt, QSize,QEvent,QCoreApplication
+from PySide6.QtWidgets import (QApplication, QMainWindow, QLabel, QWidget,
+                               QMessageBox,QWidgetItem,QFormLayout,QToolButton,QSpinBox,
+                               QDoubleSpinBox,QTextEdit,QLineEdit,QColorDialog,QHBoxLayout,
+                               QSpacerItem,QSizePolicy,QFileDialog )
+from PySide6.QtGui import QIcon,QMouseEvent
 
 import language
 import config_manager
 import run_minecraft
 import mcpartlib
 import mcpartlib.mcpartio
+import mcpartlib.mcdataformat
 
 from gui.main_window import Ui_MainWindow
-from gui.setting_color_window import Ui_SettingColorWindow
-from gui.editor_area import EditorArea,TextureRenderer
-
+from gui.editor_area import EditorArea
 # 操作列表
 draw_operations = []
 # 初始化笔刷类型
@@ -35,12 +38,45 @@ image:图片
 hexagram:六芒星
 custom:自定义
 """
+# 设置默认配置文件 
+default_config="""version: 0.0.2
+settings:
+  debug: false
+  language: zh_cn
+  safeMode: true
+  showDebugInfo: false
+  recordinglog: false
+  logformat: "[%(levelname)s at %(filename)s %(asctime)s] %(message)s "
+  logfilename: "app.log"
+  loadOtherPlugins: true
+  installMinecraft: false
+  LoadPluginMain: load_plugin.main
+paths:
+  assetsDirectory: assets
+  languageDirectory: assets/language
+  objModelDirectory: assets/obj
+  particlesConfigJson: assets/particles.json
+  particlesJsonDirectory: assets/particles
+  particlesTexturesDirectory: assets/particle
+  specialParticlesConfigJson: assets/special_particles.json
+  pluginDirectory: plugins
+"""
 # 读取配置文件
-config = config_manager.YamlFileManager("config.yaml")
+if os.path.exists("./config.yaml"):
+    if os.path.isfile("./config.yaml"):
+        config = config_manager.YamlFileManager("config.yaml")
+    else:
+        config = config_manager.YamlFileManager("config.yaml",yaml_data=default_config)
+        config.save()
+else:
+    config = config_manager.YamlFileManager("config.yaml",yaml_data=default_config)
+    config.save()
 __version__ = config['version']
+debug = config['settings']['debug']
 # 读取语言文件
-
-language_data = language.JsonFileReader(str(Path(config["paths"]["languageDirectory"]) / (config['settings']['language'] + '.json')))
+languagepath:Path=Path(config["paths"]["languageDirectory"]) / (config['settings']['language'] + '.json')
+if languagepath.exists() and languagepath.is_file():
+    language_data = language.JsonFileReader(languagepath)
 language_data.read_json()
 
 
@@ -98,55 +134,16 @@ class ErrorWindow(QWidget):
         msg.setStandardButtons(QMessageBox.Ok)
         # 显示消息框
         msg.exec()
-class SettingColorWindow(QWidget):
-    def __init__(self,parent):
-        super().__init__()
-        self.ui=Ui_SettingColorWindow()
-        self.parent=parent
-        # 创建设置颜色窗口对象
-        self.ui.setupUi(self)  # 设置 UI
-        # 更新rbga数值
-        self.update_color_value()
-    def open_window(self,/,on_window_close=lambda:None):
-        # 禁用主窗口的交互
-        self.parent.setEnabled(False)
-        # 设置该窗口为模态窗口，阻止用户与主窗口的交互
-        self.setWindowModality(Qt.ApplicationModal)
-        # 显示设置颜色窗口
-        self.show()
-        self.closeEvent=lambda event:(self.parent.setEnabled(True),self.update_color_value(),event.accept(),on_window_close())
-    def update_color_value(self) -> None:
-        self.red=self.ui.RedSpinBox.value()
-        self.green=self.ui.GreenSpinBox.value()
-        self.blue=self.ui.BlueSpinBox.value()
-        self.alpha=1#self.ui.AlphaSpinBox.value()
-    def get_color_value_rgba(self):
-        self.update_color_value()
-        return self.red,self.green,self.blue,self.alpha
-    def get_color_value_rgb(self):
-        self.update_color_value()
-        return self.red,self.green,self.blue
-    def to_qcolor(self):
-        self.update_color_value()
-        return QColor(self.red,self.green,self.blue,self.alpha)
-    def rgba_to_float(self,rgba_tuple:tuple):
-        return tuple(i / 255 for i in rgba_tuple)
-    def get_color_value_rgba_float(self):
-        self.update_color_value()
-        return self.rgba_to_float(self.get_color_value_rgba())
-    def get_color_value_rgb_float(self):
-        self.update_color_value()
-        return self.rgba_to_float(self.get_color_value_rgb())
 class MainWindow(QMainWindow):
     def __init__(self):
         logging.info(language_data.translate("main.MainWindow.logging.info.init"))
         super().__init__()
-        self.McParticleIO=mcpartlib.mcpartio.McParticleIO()
+        self.McParticleIO:None|mcpartlib.mcpartio.McParticleIO=None
         self.ui=Ui_MainWindow()
         self.ui.setupUi(self)
         # 将EditorArea替换为自定义的EditorArea
-        self.editorArea=EditorArea(self)
-        self.editorArea.setcallback(callback_mouseMoveEvent=self.record_particle,callback_paintGL=self.draw_particle)
+        self.editorArea=EditorArea(self.ui.centralwidget)
+        self.editorArea.setcallback(callback_mouse_move=self.record_particle,callback_paintGL=self.draw_particle)
         self.ui.horizontalLayout.replaceWidget(self.ui.editorArea,self.editorArea)
         
         # 读取粒子数据文件路径
@@ -172,8 +169,8 @@ class MainWindow(QMainWindow):
         asyncio.run(self.setSelectParticle())
         # -----待办----- 部分链接的功能未实现 | 优先级高
         # 链接实际功能
-        self.ui.new_N.triggered.connect(self.McParticleIO._new_file)
-        self.ui.open_O.triggered.connect(self.McParticleIO._open_file)
+        self.ui.new_N.triggered.connect(self.newFile)
+        self.ui.open_O.triggered.connect(self.openFile)
         self.ui.close_C.triggered.connect(self.closeEvent)
         self.ui.quit_Q.triggered.connect(self.close)
         self.ui.preview_P.triggered.connect(lambda : threading.Thread(target=self.preview).start())
@@ -201,8 +198,20 @@ class MainWindow(QMainWindow):
             cnt+=1
             
         logging.info(language_data.translate("main.MainWindow.logging.info.add_item",(cnt,)))
+    def newFile(self):
+        # 新建文件
+        file,_ = QFileDialog().getSaveFileName(self,filter="Minecraft Particle Datapack (*.mcpd)")	
+        if file != "":
+            logging.info(language_data.translate("main.MainWindow.logging.info.new_file",(file,)))
+            self.McParticleIO=mcpartlib.mcpartio.McParticleIO(file)
+    def openFile(self):
+        # 选择文件
+        file,_ = QFileDialog().getOpenFileName(self,filter="Minecraft Particle Datapack (*.mcpd)")
+        if file != "":
+            logging.info(language_data.translate("main.MainWindow.logging.info.open_file",(file,)))
+            self.McParticleIO=mcpartlib.mcpartio.McParticleIO(file)
     def closeEvent(self, event):
-        if self.McParticleIO.data is not None:
+        if self.McParticleIO is not None:
             # 弹出提示框询问用户是否保存文件
             reply = QMessageBox.question(self, language_data.translate("main.MainWindow.ui.text.quit"),
                                          language_data.translate("main.MainWindow.ui.text.is_save_file_needed"), 
@@ -213,10 +222,8 @@ class MainWindow(QMainWindow):
             logging.debug(language_data.translate("main.MainWindow.logging.debug.reply",(reply,event)))
             if reply == QMessageBox.StandardButton.Yes:
                 # 保存文件
-                if self.McParticleIO.filepath is None:
-                    # 如果文件路径为空，说明是新建的文件，需要另存为
-                    self.McParticleIO.save_file_as()
-                self.McParticleIO._close_file()
+                self.McParticleIO.close()
+                self.McParticleIO=None
             # 检查 event 是否是 QEvent 类型
             if not isinstance(event, QEvent):
                 logging.warning(language_data.translate("main.MainWindow.logging.warning.unexpected_event",(type(event).__name__),))
@@ -228,6 +235,7 @@ class MainWindow(QMainWindow):
                 logging.info(language_data.translate("main.MainWindow.logging.info.close"))
                 event.accept()  # 允许关闭窗口
             else:
+                logging.info(language_data.translate("main.MainWindow.logging.info.user_cancel"))
                 # 用户选择了“取消”，不关闭窗口
                 event.ignore()  # 阻止关闭窗口
     def on_select_particle_change(self):
@@ -250,6 +258,11 @@ class MainWindow(QMainWindow):
                            lambda item:item !=self.ui.SpecialOptionText and item !=self.ui.verticalSpacer)# 清空self.ui.SpecialOption中的组件
             self.replacement_values = []
             insert_cnt = 1
+            # 设置增加介绍文本函数
+            def add_introduction_text(text:str):
+                introduction_text = QLabel(self.ui.verticalLayoutWidget)
+                introduction_text.setText(text)
+                return introduction_text
             for i in particle_option_matches:
                 particle_option = i.split(":")
                 particle_option_type=particle_option[0]
@@ -259,29 +272,54 @@ class MainWindow(QMainWindow):
                         break
                 logging.debug(f"option_type: {particle_option_type}, option_name: {particle_option_introduction}")
                 if particle_option_type=="rgb" or particle_option_type=="rgba":
-                    particle_color_setting=SettingColorWindow(self)
                     # 增加介绍文本
-                    introduction_text = QLabel(self.ui.verticalLayoutWidget)
-                    introduction_text.setText(particle_option_introduction)
-                    
-                    self.ui.SpecialOption.insertWidget(insert_cnt,introduction_text)
+                    self.ui.SpecialOption.insertWidget(insert_cnt,add_introduction_text(particle_option_introduction))
                     insert_cnt+=1
                     # 增加选择颜色按钮
+                    particle_color_setting=QColorDialog()
                     color_button = QToolButton(self.ui.verticalLayoutWidget)
                     color_button.setStyleSheet(u"background-color: rgba(255, 255, 255,1);\n")
                     color_button.setAutoRaise(True)
+                    if particle_option_type=="rgba":
+                        setting_window=lambda:(particle_color_setting.getColor(options=QColorDialog.ColorDialogOption.ShowAlphaChannel),updateStyleSheet())
+                    else:
+                        setting_window=lambda:(particle_color_setting.getColor(),updateStyleSheet())
                     # 将点击信号链接到open_setting_color_window函数
                     updateStyleSheet=lambda:color_button.setStyleSheet(f"background-color: rgba{particle_color_setting.get_color_value_rgba()};\n")
-                    color_button.clicked.connect(lambda:(particle_color_setting.open_window(on_window_close=updateStyleSheet)))
+                    color_button.clicked.connect(setting_window)
                     self.ui.SpecialOption.insertWidget(insert_cnt,color_button)
                     insert_cnt+=1
                     # 根据需求设置返回
                     if particle_option_type=="rgb":
-                        self.replacement_values.append(lambda:particle_color_setting.get_color_value_rgb_float())
+                        self.replacement_values.append(lambda:particle_color_setting.getRgbF()[:3])
                     else :
-                        self.replacement_values.append(lambda:particle_color_setting.get_color_value_rgba_float())
+                        self.replacement_values.append(lambda:particle_color_setting.getRgbF())
                 elif particle_option_type=="pos":
-                    pass
+                    # 增加介绍文本
+                    self.ui.SpecialOption.insertWidget(insert_cnt,add_introduction_text(particle_option_introduction))
+                    insert_cnt+=1
+                    # 加入坐标选框
+                    pos = QHBoxLayout()
+                    x = QLabel()
+                    x.setText("x")
+                    y = QLabel()
+                    y.setText("y")
+                    z = QLabel()
+                    z.setText("z")
+                    x_spin = QDoubleSpinBox()
+                    y_spin = QDoubleSpinBox()
+                    z_spin = QDoubleSpinBox()
+                    pos.addWidget(x)
+                    pos.addWidget(x_spin)
+                    pos.addWidget(y)
+                    pos.addWidget(y_spin)
+                    pos.addWidget(z)
+                    pos.addWidget(z_spin)
+                    pos_spacer = QSpacerItem(40, 20, QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Minimum)
+                    pos.addItem(pos_spacer)
+                    self.ui.SpecialOption.insertLayout(insert_cnt,pos)
+                    insert_cnt+=1
+                    self.replacement_values.append(lambda:[x_spin.value(),y_spin.value(),z_spin.value()])
                 elif particle_option_type=="int" or particle_option_type=="float":
                     match = re.match(r"([a-zA-Z0-9]+)\[(.*)\]", particle_option_introduction)
                     if match is None:
@@ -292,28 +330,41 @@ class MainWindow(QMainWindow):
                     value_range = [float(num) for num in match.group(2).split(',')]
                     logging.debug(language_data.translate("main.MainWindow.logging.debug.print_range",(*value_range,)))
                     # 增加介绍文本
-                    introduction_text = QLabel(self.ui.verticalLayoutWidget)
-                    introduction_text.setText(particle_option_introduction)
-                    
-                    self.ui.SpecialOption.insertWidget(insert_cnt,introduction_text)
+                    self.ui.SpecialOption.insertWidget(insert_cnt,add_introduction_text(particle_option_introduction))
                     insert_cnt+=1
                     # 增加数值填写框
                     if particle_option_type=="int":
                         valueSpinBox = QSpinBox(self.ui.verticalLayoutWidget)
                     else:
                         valueSpinBox = QDoubleSpinBox(self.ui.verticalLayoutWidget)
-                    valueSpinBox.setObjectName(u"BlueSpinBox")
                     valueSpinBox.setMinimum(value_range[0])
                     valueSpinBox.setMaximum(value_range[1])
                     self.ui.SpecialOption.insertWidget(insert_cnt,valueSpinBox)
                     insert_cnt+=1
                     self.replacement_values.append(lambda:valueSpinBox.value())
                 elif particle_option_type=="snbt":
-                    pass
-                elif particle_option_type=="block":
-                    pass
-                elif particle_option_type=="item":
-                    pass
+                    # 增加介绍文本
+                    self.ui.SpecialOption.insertWidget(insert_cnt,add_introduction_text(particle_option_introduction))
+                    insert_cnt+=1
+                    # 增加输入框
+                    textEdit = QTextEdit(self.ui.verticalLayoutWidget)
+                    textEdit.setPlaceholderText(language_data.translate("main.MainWindow.ui.text.snbt_placeholder"))
+                    
+                    self.ui.SpecialOption.insertWidget(insert_cnt,textEdit)
+                    insert_cnt+=1
+                    self.replacement_values.append(lambda:textEdit.toPlainText().replace("\n",""))
+                elif particle_option_type=="block" or particle_option_type=="entity" or particle_option_type=="item":
+                    # -----待办----- 改为选择框而非手动输入,以防用户错误的输入 | 优先级中低
+                    # 增加介绍文本
+                    self.ui.SpecialOption.insertWidget(insert_cnt,add_introduction_text(particle_option_introduction))
+                    insert_cnt+=1
+                    # 增加输入框
+                    lineEdit = QLineEdit(self.ui.verticalLayoutWidget)
+                    lineEdit.setPlaceholderText(language_data.translate("main.MainWindow.ui.text.block_placeholder"))
+                    self.ui.SpecialOption.insertWidget(insert_cnt,lineEdit)
+                    insert_cnt+=1
+                    self.replacement_values.append(lambda:lineEdit.text())
+                    # -----待办结束----- 
                 else:
                     logging.error(language_data.translate("main.MainWindow.logging.error.unknown_type",(particle_option_type,)))
                     break
@@ -324,9 +375,12 @@ class MainWindow(QMainWindow):
             # 非特殊粒子就隐藏特殊属性表单
             hide_form(self.ui.SpecialOption)
     def preview(self):
-        logging.info(f"Start minecraft")
+        logging.info(language_data.translate_no_cache("main.MainWindow.logging.info.preview"))
+        
         # -----待办----- 实现数据包 | 优先级高
-
+        if self.McParticleIO is None:
+            QMessageBox.information
+            return False
         # -----待办结束----- 
         mc=run_minecraft.RunMinecraft(output_processing=lambda line,
                                       process: process.terminate() if "All chunks are save" in line else None ,
@@ -334,22 +388,17 @@ class MainWindow(QMainWindow):
         mc.install_minecraft()
         mc.run_minecraft()
         logging.info(language_data.translate_no_cache("main.MainWindow.logging.info.minecraft_exit"))
-    def record_particle(self, event: QMouseEvent):
-        return
+    def record_particle(self,EditSelf:EditorArea, event: QMouseEvent,dx,dy):
         # 获取选中的文本
         selected_text = self.ui.SelectParticle.currentText()
         # 获取选中的id
         selected_particle_id = self.particles_data[selected_text]['id']
-        if event.buttons() & Qt.RightButton:
-            draw_operation = {}
-            draw_pos = (0,0,0)#self.editorArea.screen_to_world(event.position().x(), event.position().y())
-            logging.debug(f"Mouse position: ({event.position().x()}, {event.position().y()}) -> World position: {draw_pos}")
-            draw_operation["pos"] = draw_pos
-            draw_operation["brush_type"] = brush_type
-            draw_operation["id"] = selected_particle_id
-            draw_operation["option"] = [i() for i in self.replacement_values]
-            draw_operations.append(draw_operation)
-
+        if  Qt.MouseButton.RightButton in EditSelf.mouseButtons:
+            draw_pos = tuple(EditSelf.click_pos)
+            option = [i() for i in self.replacement_values]
+            particle_draw=mcpartlib.mcdataformat.McParticleData(_type=brush_type,_particle_id=selected_particle_id,_pos=draw_pos,_option=option)
+            draw_operations.append(particle_draw)
+            self.McParticleIO._add_data(particle_draw)
     def draw_particle(self):
         return
         for i in draw_operations:
@@ -367,6 +416,7 @@ if __name__ == "__main__":
     try :
         window = MainWindow()
         window.show()
+        logging.info(language_data.translate_no_cache("main.top.logging.info.init",(time.perf_counter() - start_time,)))
         exit_code = app.exec()
     except RuntimeError as e:
         logging.error(e)
